pre DB switch changes

+++ TyperV1API/Controllers/UserController.cs
@@ -94,35 +94,31 @@ namespace TyperV1API.Controllers
                 return BadRequest(u.UserName + " is already in use");
             }
 
-            User newUser = new User();
-
-            newUser.Joined = u.Joined;
-            newUser.UserName = u.UserName;
-            newUser.Password = passwordService.HashPassword(u.Password);
-            newUser.FirstName = u.FirstName;
-            newUser.LastName = u.LastName;
-            newUser.Email = u.Email;
+            User newUser = new User
+            {
+                UserName = u.UserName,
+                Password = passwordService.HashPassword(u.Password),
+                FirstName = u.FirstName,
+                LastName = u.LastName,
+                Email = u.Email
+            };
 
             if (u.Image != null)
             {
                 Image newImage = uploader.getImage(u.Image, "Users");
                 if (newImage != null)
                 {
+                    dbContext.Images.Add(newImage);
+                    await dbContext.SaveChangesAsync();
+
                     newUser.ImageId = newImage.ImageId;
-                    newUser.Image = await dbContext.Images.FindAsync(newUser.ImageId);
                 }
-            }
-            else
-            {
-                newUser.ImageId = 1;
-                newUser.Image = await dbContext.Images.FindAsync(newUser.ImageId);
             }
 
             dbContext.Users.Add(newUser);
             await dbContext.SaveChangesAsync();
 
             return Ok(convertUserDTO(newUser));
-
         }
 
         [HttpPut("{id}")]
@@ -172,6 +168,11 @@ namespace TyperV1API.Controllers
                     updateUser.Image = await dbContext.Images.FindAsync(updateUser.ImageId);
                 }
             }
+            if (updateUser.ImageId == null)
+            {
+                updateUser.ImageId = 1;
+                updateUser.Image = await dbContext.Images.FindAsync(1);
+            }
 
             dbContext.Users.Update(updateUser);
             await dbContext.SaveChangesAsync();
@@ -183,7 +184,7 @@ namespace TyperV1API.Controllers
         [HttpDelete("{id}")]
         public async Task<IActionResult> removeUser(int id)
         {
-            User result = await dbContext.Users.FindAsync(id);
+            User result = await dbContext.Users.Include(i => i.Image).FirstOrDefaultAsync(u => u.UserId == id);
 
             if (result == null || result.Active == false)
             {
@@ -192,13 +193,18 @@ namespace TyperV1API.Controllers
 
             result.Active = false;
 
-            if(result.Image != null)
+            if (result.Image != null && System.IO.File.Exists(Path.Combine(Directory.GetCurrentDirectory(), result.Image.ImagePath))
+                && result.Image.ImagePath != "Images\\DefaultProfPic\\V1DefaultProfPic.webp")
             {
                 System.IO.File.Delete(Path.Combine(Directory.GetCurrentDirectory(), result.Image.ImagePath));
+                dbContext.Images.Remove(result.Image);
             }
 
-            result.ImageId = 1;
-            result.Image = await dbContext.Images.FindAsync(result.ImageId);
+            if (result.ImageId == null)
+            {
+                result.ImageId = 1;
+                result.Image = await dbContext.Images.FindAsync(1);
+            }
 
             dbContext.Users.Update(result);
             await dbContext.SaveChangesAsync();
